+++ src/components/VenueList/index.jsx
@@ -1,9 +1,9 @@
+import VenueCard from "../VenueCard";
 import Grid from "@mui/material/Grid";
 import Typography from "@mui/material/Typography";
-import VenueCard from "../VenueCard";
 
+// This component is used to display a list of venues
 const VenueList = ({ venues }) => {
-  console.log(venues);
   return (
     <Grid container spacing={4}>
       {venues.length === 0 ? (
