+++ src/pages/Home/index.jsx
@@ -1,26 +1,25 @@
 import { useEffect } from "react";
-import useAuthStore from "../../stores/authStore";
-import useVenueStore from "../../stores/venueStore";
+import { useVenueStore, useFetchStore } from "../../stores";
 import VenueList from "../../components/VenueList";
 import Container from "@mui/material/Container";
 
-// Home page component
+// This component is used to display a home page with a list of venues
 const Home = () => {
-  // Get token from authStore
-  const { token } = useAuthStore();
   // Get states and actions from venuesStore
-  const { venues, isLoading, isError, fetchVenues } = useVenueStore();
+  const fetchVenues = useVenueStore((state) => state.fetchVenues);
+  const isLoading = useFetchStore((state) => state.isLoading);
+  const isError = useFetchStore((state) => state.isError);
+  const errorMsg = useFetchStore((state) => state.errorMsg);
+  const venues = useVenueStore((state) => state.venues);
 
   // Fetch venues when token and fetchVenues function change
   useEffect(() => {
-    fetchVenues(token);
-  }, [token, fetchVenues]);
+    fetchVenues();
+  }, []);
 
   if (isLoading) return <h1>Loading...</h1>;
 
-  if (isError) return <h1>Error</h1>;
-
-  console.log(venues);
+  if (isError) return <h1>Error: {errorMsg}</h1>;
 
   return (
     <Container>
